DNSExplorer can now enumerate web servers, gathering basic information by using whatweb, the script is able to determine data such as IP address, web application programming language, web server, and can detect if a WAF is protecting the website by wafw00f.
This adds new dependencies, such as whatweb and wafw00f, which are already installed with the default kali-linux configuration, however, in case you don't have them, their installation is easy and they are available for most linux distributions.

Another feature is that DNSExplorer now exports a CSV file that condenses all the web server enumeration information, giving an overview of the target organization's web application firewall technologies and security posture, all at a glance.

Now DNSExplorer uses dynamic threads for the execution of its most tedious tasks, which has exponentially improved its performance, to give an idea of this, the first functional version of DNSExplorer (v1.5.0) took 3.4 minutes to consume the default dictionary for the subdomain discovery attack, while this new version takes only 21 seconds, which represents an increase of approximately 89.71% in the performance of this task, using only 10 threads. However, DNSExplorer dynamically calculates the best distribution of threads for its tasks, some heavy tasks such as web server discovery, server technology enumeration or WAF discovery can take more than 15 seconds to complete due to the timeout of each of these tools, including other variables such as connection quality and server response time. DNSExplorer includes a range from 1 to 25 threads for the most demanding tasks, using the number of sites to enumerate as the main variable to define the number of threads.
Therefore, if we now use a custom dictionary with few lines it will be consumed quickly optimizing processor resources in a single sequential execution, but, if we use a dictionary containing, for example, 100,000 or even more, we can take advantage of the full power of the processor to complete multiple tasks simultaneously, something impossible in the previous version.